Updated syntax of logic to be more free flowing.  We no longer use | to distinguish between pure and spatial assertions.  Instead infix operators are assumed to be pure and normal predicates are assumed to be spatial.
Only seemed to really affect

  type(x,C)

needs replacing with 
  x : C

We can now write 
  x < y * y < z || z < y * field(g,f,x) * y < x  
which would previously have been written

|  (LT(x,y) * LT(y,z) | ) 
  || (LT(z,y) * LT(y,x) | field(g,f,x)) 

We can also write infix operators in expressions such as

  (x+y)=(y+x)

I have updated the examples directory to use these notation.  We
output a deprecated warning for use of | in specifications, I will
remove support in due course to make the parsing simpler.

Also, I still need to allow this syntax in the logic files, which
involves a reasonable amount of merging of the lex and Yac files.
Probably will be done in the next couple of days.