Lots of small code rewrites.
For example, strictly speaking OCaml does not guarantee that all hell
doesn't break loose when Format and Printf calls are mixed, which means
that it's better to stick with Format. Once you do that, then it's
natural to use its tagging mechanism for the terminal color changes,
instead of the old system from system.ml. Once you rewrite the code that
prints colored stuff you bump into lots of places that should use the
logging infrastructure that I commited last time, and so on.

Another change that may warant mention here is the module
Misc.MapHelper.  See its comments.