A dumb code generator for `evaluators'.
An evaluator is an object that takes some complicated tree-like
data structure and computes a value. The generated code takes care
of the tree traversal. This is useful when the tree is fairly
heterogeneous (there are many "type t =  ..." invlved), but the
value depends only of few parts of it (with certain types).

For example, the data types used for representing a jimple file are
fairly complex. If you want to extract all the class names mentioned
in a jimple file you need to write lots of uninteresting traversal code
... or subclass a generated jimple evaluator.